state value should be at most 64bit

+++ src/main/java/net/goldenstack/minestom_ca/AutomataQuery.java
@@ -3,13 +3,13 @@ package net.goldenstack.minestom_ca;
 import java.util.Map;
 
 public interface AutomataQuery {
-    int stateAt(int x, int y, int z, int index);
+    long stateAt(int x, int y, int z, int index);
 
-    default int blockAt(int x, int y, int z) {
+    default long blockAt(int x, int y, int z) {
         return stateAt(x, y, z, 0);
     }
 
-    Map<Integer, Integer> queryIndexes(int x, int y, int z);
+    Map<Integer, Long> queryIndexes(int x, int y, int z);
 
-    Map<String, Integer> queryNames(int x, int y, int z);
+    Map<String, Long> queryNames(int x, int y, int z);
 }
